fix n79 unit test

+++ lib/include/srsran/common/band_helper.h
@@ -136,7 +136,7 @@ public:
    */
   static srsran_ssb_pattern_t get_ssb_pattern(uint16_t band, srsran_subcarrier_spacing_t scs);
 
-  static uint32_t get_freq_from_gscn(uint32_t gscn);
+  static uint64_t get_freq_from_gscn(uint32_t gscn);
 
   /**
    * @brief Select the lower SSB subcarrier spacing valid for this band
